refactor(testdriver-api): add record as default implementation
- readd failureCount() method which is used by slack notification
- move TestReportImpl to where it is being used
